# _reversa_sdd/openapi/brolabtask-api.yaml
openapi: "3.0.3"
info:
  title: BrolabTask API
  version: "2.0"
  description: |
    API REST do BrolabTask — Kanban CLI-themed. Todos os endpoints são Route Handlers do Next.js App Router.
    Base URL: /api
    Autenticação: **Nenhuma** (todos os endpoints são públicos) 🔴 CRÍTICO
servers:
  - url: /api
paths:
...
  /tasks:
    get:
      summary: Listar todas as tarefas
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
    post:
      summary: Criar tarefa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, columnId]
              properties:
                title:
                  type: string
                description:
                  type: string
                columnId:
                  type: string
                  format: uuid
                position:
                  type: integer
                assignees:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
    patch:
      summary: Atualizar tarefa (incluindo mover entre colunas)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: string
                  format: uuid
                title:
                  type: string
                description:
                  type: string
                columnId:
                  type: string
                  format: uuid
                position:
                  type: integer
                assignees:
                  type: array
                  items:
                    type: string
                labels:
                  type: array
                  items:
                    $ref: "#/components/schemas/Label"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
    delete:
      summary: Deletar tarefa e seus arquivos
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        columnId:
          type: string
          format: uuid
        position:
          type: integer
        assignees:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        files:
          type: array
          items:
            $ref: "#/components/schemas/TaskFile"
        createdAt:
          type: string
          format: date-time
...
    Label:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authorId:
          type: string
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        mentions:
          type: array
          items:
            type: string
    TaskFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        size:
          type: integer
          description: Tamanho em bytes
        type:
          type: string
        url:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
...